generator works
this generator seems to work properly. We still have to test it but so far, looks like it works (at least prints as it should). There's a little thing to correct that it's that can generate more than one test at once but...it's fine once we solve that. 
We could also add c++11 to use to_string